split class into separate file
